+++ lib/animated_textformfields.dart
@@ -1,3 +1,5 @@
 library animated_textformfields;
 
+export 'definitions/callbacks.dart';
 export 'animated_textformfield/animated_textformfield.dart';
+export 'custom_textformfield_lite/custom_textformfield_lite.dart';