+++ lib/definitions/callbacks.dart
@@ -1,0 +1,2 @@
+// defining a stringcallback that will be used to display validation messages
+typedef StringCallback = String Function(String);